auth/requirements.txt
aiokafka==0.8.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
black==23.7.0
click==8.1.6
ecdsa==0.18.0
fastapi==0.101.0
h11==0.14.0
idna==3.4
kafka-python==2.0.2
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
peewee==3.16.2
platformdirs==3.10.0
pyasn1==0.5.0
pydantic==2.1.1
pydantic_core==2.4.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
